backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.36
alembic==1.13.3
psycopg[binary]==3.2.3
python-dotenv==1.0.1
pandas==3.0.1
usaddress==0.5.10
rapidfuzz==3.9.7
pytest==8.3.3
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-dateutil==2.9.0
schedule==1.2.2